refactor: consolidate history reader
